refactor(adapters): Created folders for each adapters to contain adapter file and tests files.
