+++ packages/api-gw/src/adapters/hash/hash.js
@@ -1,16 +1,4 @@
-const { createError } = require('../utils/errorHandling');
-
-async function createShortUrl(req, reply) {
-  const { url } = req.body;
-  
-  req.log.info(`Creating short url for ${url}`);
-  const hostnameUrl = `${req.protocol}://${req.hostname}`;
-  const shortUrL = await this.linkAdapter.createShortUrl(url, hostnameUrl);
-
-  return reply.code(201)
-    .headers('Content-Type', 'application/json; charset=utf-8')
-    .send({ data: shortUrL });
-}
+const { createError } = require('../../utils/errorHandling');
 
 async function redirectToUrl(req, reply) {
   const { hash } = req.params;
@@ -31,6 +19,5 @@ async function redirectToUrl(req, reply) {
 }
 
 module.exports = {
-  createShortUrl,
   redirectToUrl
 };
